implement wasm by means of a separate module
This is the approach currently in use in core-crypto, and is known
to work. It certainly simplifies things! We'd really like to avoid
it, though: it doubles the effective surface area of the code, making
maintenance much more difficult.

That said, there was some effort expended (without success) to using
the single-source model previously. So I'm committing this work for
now, as a reference point.